simplify netsuite index, split files
